src/test/java/fim/unipassau/de/scratchLog/web/DashboardRestControllerTest.java
/*
 * Copyright (C) 2023 ScratchLog contributors
 *
 * This file is part of ScratchLog.
 *
 * ScratchLog is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or (at
 * your option) any later version.
 *
 * ScratchLog is distributed in the hope that it will be useful, but
 * WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
 * General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with ScratchLog. If not, see <http://www.gnu.org/licenses/>.
 */

package fim.unipassau.de.scratchLog.web;

import fim.unipassau.de.scratchLog.application.service.DashboardService;
import fim.unipassau.de.scratchLog.web.controller.DashboardRestController;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.mockito.ArgumentMatchers.anyInt;
import static org.mockito.Mockito.never;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
public class DashboardRestControllerTest {

    @InjectMocks
    private DashboardRestController dashboardRestController;

    @Mock
    private DashboardService dashboardService;

    private static final String ID_STRING = "5";
    private static final int ID = 5;
    private static final String[] experimentData = new String[]{"11", "7", "5"};

    @Test
    public void testGetExperimentData() {
        when(dashboardService.getExperimentData(ID)).thenReturn(experimentData);
        assertEquals(experimentData, dashboardRestController.getExperimentData(ID_STRING));
        verify(dashboardService).getExperimentData(ID);
    }

    @Test
    public void testGetExperimentDataInvalidId() {
        assertThrows(IllegalArgumentException.class,
                () -> dashboardRestController.getExperimentData(null)
        );
        verify(dashboardService, never()).getExperimentData(anyInt());
    }

}
